fixes #3
@joeshaw spotted a bug where sending a GET request via the OAuth 1.0/a hook
that contained URL parameters would cause duplication of the params in the
final URL rendering. This turned out to be related to the fact that
request.params is precalculated prior to entering the pre-request hook.
Initially it looked like there was no clean solution to this problem, however
the Request object provides an internal method, `_encode_params` which allowed
for a simple enough hack.

However there remains a persistent issue with request bodies which suffer from
a similar problem but cannot be overriden in the scope of the hook. This is a
known bug and there is at least one proposed solution. One salient problem
persists: due to planned streaming support, the proposed solution cannot be
easily merged into Requests. It's unclear when this problem will be patched.