updating to 0.3.0, fixes #9
This release is not directly compatiable with the previous releases as the API
has been updated and extended. Namely the service wrapper request calls have
been made more generic to allow for various HTTP methods as well as flexibility
when indicating the name of provider credentials, such as access_token.

Specifically this addresses an issue where a service uses a non-standard naming
scheme, such as Foursquare. Previously there was no easy way to overide the
parameter name and internally the variable was required or the method would
throw an error. Now the method has been abstracted to allow for generic naming
as well as dynamic HTTP methods.

Also the API has been extended and modified to be more closely aligned with
Request's. For instance, all service wrappers now have get(), post(), put(),
and delete() methods and corrospond to their respective HTTP methods. Some
naming conventions were altered, such as the http_method parameter which is now
simply method. (This also mirrors Request's API.)