fixing a bug where params containing whitespace are not properly encoded
Previously parameters that contained whitespace, while quoted and encoded, were
not escaped. This would cause authentication problems due to the fact that
the signature being generated was incorrect. Now this is solved by calling
replace('+', '%20') on the string returned by urlencode.

Additionally handling of OAuth 1.0/a POSTs has been updated to include the
oauth paramters in the query string. This is a stopgap fix, the fact that the
request body cannot be updated within the scope of a hook remains an issue on
Requests. This will be updated once a fix has been pushed there.