adding a LinkedIn example app, fixes #11
This updates rauth to v0.4.1 and fixes a bug wherein POSTing to a given
endpoint could not contain the oauth_verifier parameter. Now instead of
keeping a record of the verifier on the token object we assign it as a
property of the hook. This property is populated automatically: in the scope
of the hook's `__call__` method we introspect URL parameters and data, if we
find `oauth_verifier` in either this value is assigned to our hook instance
as a property. Otherwise this can be manually overridden by explicit
assignment.