updating to address an issue with header authentication
Essentially header authenication would fail under certain circumstances due to
the fact that optional oauth parameters, such as `oauth_verifier` were being
appended multiple times. This is now solved by parsing out these paramters
before generating the `oauth_params` property.

Additionally Token objects have been removed; they add unneeded complexity to
the wrapper. Values of tokens are simply set as attributes directly on the hook
itself. This simplifies this portion of the library and reduces the number of
classes needed without breaking existing functionality.

Finally `get_authenticated_session` has been removed and replaced by
`_get_session`. This also signals an important aspect of this update: state
of the hook was not being saved between authentication steps or even calls.
Now through the use of this new method and instantiation of the hook as an
attribute of the service wrapper, we are able to update attribtes on the
hook object, rather than re-instantiate it on each call.