updating OAuth 1.0 entity-method handling to reflect spec
Previously OAuth 1.0/a was improperly implemented causing non-urlencoded body
data to be lost or worse added to the base signature string. According to the
OAuth sopecification, there are strict requirements[1] for signing against
entity-bodies.

Now we bring the `OAuth1Session` object closer to the specification
requirements by only adding the entity-body to the base string signature when
we are handed an explicit Content-Type of `x-www-form-urlencoded` and have
a so-called "entity" method, i.e. one of POST, PUT, PATCH.

[1] http://tools.ietf.org/html/rfc5849#section-3.5.2

fixes: #68